Temporary workaround script for testing all mergeable pull requests.
The script uses the github api to check wich pr are mergeable and not
yet tested. Then it runs sympy-bot on them.

The solution is very ugly. The way to use the script is just to start it
from your sympy-bot directory.

You _must_ have a .sympy-bot.conf file for this to work.