Retry when it could not get the pull request
This also changes all retries to use a doubled waiting scheme, where it waits
for 1 second, then 2 seconds, then 4 seconds, then 8 seconds, and so on.  This
is for example what GMail does.  This prevents us from DOSing the server.

The easiest way to test this is to disconnect yourself from the internet and
try to review a pull request.

Closes #96.