Add support for building docs
The bot can now build the docs, using the default 'make html-errors'
command, launched by default from the 'doc' directory. This behavior can
be toggled on with the '-D' option. The command to build the docs and
the directory from which to build the docs can both be configured from
command line options.