Handle the case where a pull request is really an issue
In this case, a message is printed, and the pull request is skipped.  We skip
rather than die in this case so that we can gracefully move on to the next
pull request in the list.  This way, things like sympy-bot review 1000 1001
1002 1003 ... will just continue to the next pull request if one of them is
just an issue.

If a pull request is an issue, the pulls api gives 404, but the issues api
returns, so if we get 404, we check that, and if it exists, we conclude that
the pull request must really be an issue.

This closes #143.