Factor out time doubling retry logic into a single function, keep_trying()
See the docstring of keep_trying() for more information.  The basic idea is
that the function tries something, and if one of the given exceptions is
raised, it prints an error message, sleeps, and then tries again.  The time
slept doubles with each failure, in order to prevent DoS.  It also supports
doing additional error handling when the error is caught.

I also changed instances of catching HTTPError to just URLError, since the
former is a subclass of the latter.